Move more solvers, try to resolve some testing issues
