+++ HARK/ConsumptionSaving/ConsMedModel.py
@@ -46,7 +46,6 @@ __all__ = [
     "cThruXfunc",
     "MedThruXfunc",
     "MedShockConsumerType",
-    "ConsMedShockSolver",
 ]
 
 utility_inv = CRRAutility_inv
@@ -1363,653 +1362,3 @@ def solve_one_period_ConsMedShock(
     solution_now.MedFunc = MedFuncNow
     solution_now.policyFunc = policyFuncNow
     return solution_now
-
-
-class ConsMedShockSolver(ConsGenIncProcessSolver):
-    """
-    Class for solving the one period problem for the "medical shocks" model, in
-    which consumers receive shocks to permanent and transitory income as well as
-    shocks to "medical need"-- multiplicative utility shocks for a second good.
-
-    Parameters
-    ----------
-    solution_next : ConsumerSolution
-        The solution to next period's one period problem.
-    IncShkDstn : distribution.Distribution
-        A discrete
-        approximations to the income process between the period being solved
-        and the one immediately following (in solution_next).
-    MedShkDstn : distribution.Distribution
-        Discrete distribution of the multiplicative utility shifter for med-
-        ical care.
-    LivPrb : float
-        Survival probability; likelihood of being alive at the beginning of
-        the succeeding period.
-    DiscFac : float
-        Intertemporal discount factor for future utility.
-    CRRA : float
-        Coefficient of relative risk aversion for composite consumption.
-    CRRAmed : float
-        Coefficient of relative risk aversion for medical care.
-    Rfree : float
-        Risk free interest factor on end-of-period assets.
-    MedPrice : float
-        Price of unit of medical care relative to unit of consumption.
-    pLvlNextFunc : float
-        Expected permanent income next period as a function of current pLvl.
-    BoroCnstArt: float or None
-        Borrowing constraint for the minimum allowable assets to end the
-        period with.
-    aXtraGrid: np.array
-        Array of "extra" end-of-period (normalized) asset values-- assets
-        above the absolute minimum acceptable level.
-    pLvlGrid: np.array
-        Array of permanent income levels at which to solve the problem.
-    vFuncBool: boolean
-        An indicator for whether the value function should be computed and
-        included in the reported solution.
-    CubicBool: boolean
-        An indicator for whether the solver should use cubic or linear inter-
-        polation.
-    """
-
-    def __init__(
-        self,
-        solution_next,
-        IncShkDstn,
-        MedShkDstn,
-        LivPrb,
-        DiscFac,
-        CRRA,
-        CRRAmed,
-        Rfree,
-        MedPrice,
-        pLvlNextFunc,
-        BoroCnstArt,
-        aXtraGrid,
-        pLvlGrid,
-        vFuncBool,
-        CubicBool,
-    ):
-        """
-        Constructor for a new solver for a one period problem with idiosyncratic
-        shocks to permanent and transitory income and shocks to medical need.
-        """
-        self.solution_next = solution_next
-        self.IncShkDstn = IncShkDstn
-        self.MedShkDstn = MedShkDstn
-        self.LivPrb = LivPrb
-        self.DiscFac = DiscFac
-        self.CRRA = CRRA
-        self.CRRAmed = CRRAmed
-        self.Rfree = Rfree
-        self.MedPrice = MedPrice
-        self.pLvlNextFunc = pLvlNextFunc
-        self.BoroCnstArt = BoroCnstArt
-        self.aXtraGrid = aXtraGrid
-        self.pLvlGrid = pLvlGrid
-        self.vFuncBool = vFuncBool
-        self.CubicBool = CubicBool
-        self.PermGroFac = 0.0
-        self.def_utility_funcs()
-
-    def set_and_update_values(self, solution_next, IncShkDstn, LivPrb, DiscFac):
-        """
-        Unpacks some of the inputs (and calculates simple objects based on them),
-        storing the results in self for use by other methods.  These include:
-        income shocks and probabilities, medical shocks and probabilities, next
-        period's marginal value function (etc), the probability of getting the
-        worst income shock next period, the patience factor, human wealth, and
-        the bounding MPCs.
-
-        Parameters
-        ----------
-        solution_next : ConsumerSolution
-            The solution to next period's one period problem.
-        IncShkDstn : distribution.Distribution
-            A discrete
-            approximation to the income process between the period being solved
-            and the one immediately following (in solution_next).
-        LivPrb : float
-            Survival probability; likelihood of being alive at the beginning of
-            the succeeding period.
-        DiscFac : float
-            Intertemporal discount factor for future utility.
-
-        Returns
-        -------
-        None
-        """
-        # Run basic version of this method
-        ConsGenIncProcessSolver.set_and_update_values(
-            self, self.solution_next, self.IncShkDstn, self.LivPrb, self.DiscFac
-        )
-
-        # Also unpack the medical shock distribution
-        self.MedShkPrbs = self.MedShkDstn.pmv
-        self.MedShkVals = self.MedShkDstn.atoms.flatten()
-
-    def def_utility_funcs(self):
-        """
-        Defines CRRA utility function for this period (and its derivatives,
-        and their inverses), saving them as attributes of self for other methods
-        to use.  Extends version from ConsIndShock models by also defining inverse
-        marginal utility function over medical care.
-
-        Parameters
-        ----------
-        none
-
-        Returns
-        -------
-        none
-        """
-        ConsGenIncProcessSolver.def_utility_funcs(self)  # Do basic version
-        self.uMed = UtilityFuncCRRA(self.CRRAmed)
-
-    def def_BoroCnst(self, BoroCnstArt):
-        """
-        Defines the constrained portion of the consumption function as cFuncNowCnst,
-        an attribute of self.  Uses the artificial and natural borrowing constraints.
-
-        Parameters
-        ----------
-        BoroCnstArt : float or None
-            Borrowing constraint for the minimum allowable (normalized) assets
-            to end the period with.  If it is less than the natural borrowing
-            constraint at a particular permanent income level, then it is irrelevant;
-            BoroCnstArt=None indicates no artificial borrowing constraint.
-
-        Returns
-        -------
-        None
-        """
-        # Find minimum allowable end-of-period assets at each permanent income level
-        PermIncMinNext = self.PermShkMinNext * self.pLvlNextFunc(self.pLvlGrid)
-        IncLvlMinNext = PermIncMinNext * self.TranShkMinNext
-        aLvlMin = (
-            self.solution_next.mLvlMin(PermIncMinNext) - IncLvlMinNext
-        ) / self.Rfree
-
-        # Make a function for the natural borrowing constraint by permanent income
-        BoroCnstNat = LinearInterp(
-            np.insert(self.pLvlGrid, 0, 0.0), np.insert(aLvlMin, 0, 0.0)
-        )
-        self.BoroCnstNat = BoroCnstNat
-
-        # Define the minimum allowable level of market resources by permanent income
-        if self.BoroCnstArt is not None:
-            BoroCnstArt = LinearInterp([0.0, 1.0], [0.0, self.BoroCnstArt])
-            self.mLvlMinNow = UpperEnvelope(BoroCnstNat, BoroCnstArt)
-        else:
-            self.mLvlMinNow = BoroCnstNat
-
-        # Make the constrained total spending function: spend all market resources
-        trivial_grid = np.array([0.0, 1.0])  # Trivial grid
-        spendAllFunc = TrilinearInterp(
-            np.array([[[0.0, 0.0], [0.0, 0.0]], [[1.0, 1.0], [1.0, 1.0]]]),
-            trivial_grid,
-            trivial_grid,
-            trivial_grid,
-        )
-        self.xFuncNowCnst = VariableLowerBoundFunc3D(spendAllFunc, self.mLvlMinNow)
-
-        self.mNrmMinNow = (
-            0.0  # Needs to exist so as not to break when solution is created
-        )
-        self.MPCmaxEff = (
-            0.0  # Actually might vary by p, but no use formulating as a function
-        )
-
-    def get_points_for_interpolation(self, EndOfPrdvP, aLvlNow):
-        """
-        Finds endogenous interpolation points (x,m) for the expenditure function.
-
-        Parameters
-        ----------
-        EndOfPrdvP : np.array
-            Array of end-of-period marginal values.
-        aLvlNow : np.array
-            Array of end-of-period asset values that yield the marginal values
-            in EndOfPrdvP.
-
-        Returns
-        -------
-        x_for_interpolation : np.array
-            Total expenditure points for interpolation.
-        m_for_interpolation : np.array
-            Corresponding market resource points for interpolation.
-        p_for_interpolation : np.array
-            Corresponding permanent income points for interpolation.
-        """
-        # Get size of each state dimension
-        mCount = aLvlNow.shape[1]
-        pCount = aLvlNow.shape[0]
-        MedCount = self.MedShkVals.size
-
-        # Calculate endogenous gridpoints and controls
-        cLvlNow = np.tile(
-            np.reshape(self.u.derinv(EndOfPrdvP, order=(1, 0)), (1, pCount, mCount)),
-            (MedCount, 1, 1),
-        )
-        MedBaseNow = np.tile(
-            np.reshape(
-                self.uMed.derinv(self.MedPrice * EndOfPrdvP, order=(1, 0)),
-                (1, pCount, mCount),
-            ),
-            (MedCount, 1, 1),
-        )
-        MedShkVals_tiled = np.tile(
-            np.reshape(self.MedShkVals ** (1.0 / self.CRRAmed), (MedCount, 1, 1)),
-            (1, pCount, mCount),
-        )
-        MedLvlNow = MedShkVals_tiled * MedBaseNow
-        aLvlNow_tiled = np.tile(
-            np.reshape(aLvlNow, (1, pCount, mCount)), (MedCount, 1, 1)
-        )
-        xLvlNow = cLvlNow + self.MedPrice * MedLvlNow
-        mLvlNow = xLvlNow + aLvlNow_tiled
-
-        # Limiting consumption is zero as m approaches the natural borrowing constraint
-        x_for_interpolation = np.concatenate(
-            (np.zeros((MedCount, pCount, 1)), xLvlNow), axis=-1
-        )
-        temp = np.tile(
-            self.BoroCnstNat(np.reshape(self.pLvlGrid, (1, self.pLvlGrid.size, 1))),
-            (MedCount, 1, 1),
-        )
-        m_for_interpolation = np.concatenate((temp, mLvlNow), axis=-1)
-
-        # Make a 3D array of permanent income for interpolation
-        p_for_interpolation = np.tile(
-            np.reshape(self.pLvlGrid, (1, pCount, 1)), (MedCount, 1, mCount + 1)
-        )
-
-        # Store for use by cubic interpolator
-        self.cLvlNow = cLvlNow
-        self.MedLvlNow = MedLvlNow
-        self.MedShkVals_tiled = np.tile(
-            np.reshape(self.MedShkVals, (MedCount, 1, 1)), (1, pCount, mCount)
-        )
-
-        return x_for_interpolation, m_for_interpolation, p_for_interpolation
-
-    def use_points_for_interpolation(self, xLvl, mLvl, pLvl, MedShk, interpolator):
-        """
-        Constructs a basic solution for this period, including the consumption
-        function and marginal value function.
-
-        Parameters
-        ----------
-        xLvl : np.array
-            Total expenditure points for interpolation.
-        mLvl : np.array
-            Corresponding market resource points for interpolation.
-        pLvl : np.array
-            Corresponding permanent income level points for interpolation.
-        MedShk : np.array
-            Corresponding medical need shocks for interpolation.
-        interpolator : function
-            A function that constructs and returns a consumption function.
-
-        Returns
-        -------
-        solution_now : ConsumerSolution
-            The solution to this period's consumption-saving problem, with a
-            consumption function, marginal value function, and minimum m.
-        """
-        # Construct the unconstrained total expenditure function
-        xFuncNowUnc = interpolator(mLvl, pLvl, MedShk, xLvl)
-        xFuncNowCnst = self.xFuncNowCnst
-        xFuncNow = LowerEnvelope3D(xFuncNowUnc, xFuncNowCnst)
-
-        # Transform the expenditure function into policy functions for consumption and medical care
-        aug_factor = 2
-        xLvlGrid = make_grid_exp_mult(
-            np.min(xLvl), np.max(xLvl), aug_factor * self.aXtraGrid.size, 8
-        )
-        policyFuncNow = MedShockPolicyFunc(
-            xFuncNow,
-            xLvlGrid,
-            self.MedShkVals,
-            self.MedPrice,
-            self.CRRA,
-            self.CRRAmed,
-            xLvlCubicBool=self.CubicBool,
-        )
-        cFuncNow = cThruXfunc(xFuncNow, policyFuncNow.cFunc)
-        MedFuncNow = MedThruXfunc(xFuncNow, policyFuncNow.cFunc, self.MedPrice)
-
-        # Make the marginal value function (and the value function if vFuncBool=True)
-        vFuncNow, vPfuncNow = self.make_v_and_vP_funcs(policyFuncNow)
-
-        # Pack up the solution and return it
-        solution_now = ConsumerSolution(
-            cFunc=cFuncNow, vFunc=vFuncNow, vPfunc=vPfuncNow, mNrmMin=self.mNrmMinNow
-        )
-        solution_now.MedFunc = MedFuncNow
-        solution_now.policyFunc = policyFuncNow
-        return solution_now
-
-    def make_v_and_vP_funcs(self, policyFunc):
-        """
-        Constructs the marginal value function for this period.
-
-        Parameters
-        ----------
-        policyFunc : function
-            Consumption and medical care function for this period, defined over
-            market resources, permanent income level, and the medical need shock.
-
-        Returns
-        -------
-        vFunc : function
-            Value function for this period, defined over market resources and
-            permanent income.
-        vPfunc : function
-            Marginal value (of market resources) function for this period, defined
-            over market resources and permanent income.
-        """
-        # Get state dimension sizes
-        mCount = self.aXtraGrid.size
-        pCount = self.pLvlGrid.size
-        MedCount = self.MedShkVals.size
-
-        # Make temporary grids to evaluate the consumption function
-        temp_grid = np.tile(
-            np.reshape(self.aXtraGrid, (mCount, 1, 1)), (1, pCount, MedCount)
-        )
-        aMinGrid = np.tile(
-            np.reshape(self.mLvlMinNow(self.pLvlGrid), (1, pCount, 1)),
-            (mCount, 1, MedCount),
-        )
-        pGrid = np.tile(
-            np.reshape(self.pLvlGrid, (1, pCount, 1)), (mCount, 1, MedCount)
-        )
-        mGrid = temp_grid * pGrid + aMinGrid
-        if self.pLvlGrid[0] == 0:
-            mGrid[:, 0, :] = np.tile(
-                np.reshape(self.aXtraGrid, (mCount, 1)), (1, MedCount)
-            )
-        MedShkGrid = np.tile(
-            np.reshape(self.MedShkVals, (1, 1, MedCount)), (mCount, pCount, 1)
-        )
-        probsGrid = np.tile(
-            np.reshape(self.MedShkPrbs, (1, 1, MedCount)), (mCount, pCount, 1)
-        )
-
-        # Get optimal consumption (and medical care) for each state
-        cGrid, MedGrid = policyFunc(mGrid, pGrid, MedShkGrid)
-
-        # Calculate expected value by "integrating" across medical shocks
-        if self.vFuncBool:
-            MedGrid = np.maximum(
-                MedGrid, 1e-100
-            )  # interpolation error sometimes makes Med < 0 (barely)
-            aGrid = np.maximum(
-                mGrid - cGrid - self.MedPrice * MedGrid, aMinGrid
-            )  # interpolation error sometimes makes tiny violations
-            vGrid = (
-                self.u(cGrid)
-                + MedShkGrid * self.uMed(MedGrid)
-                + self.EndOfPrdvFunc(aGrid, pGrid)
-            )
-            vNow = np.sum(vGrid * probsGrid, axis=2)
-
-        # Calculate expected marginal value by "integrating" across medical shocks
-        vPgrid = self.u.der(cGrid)
-        vPnow = np.sum(vPgrid * probsGrid, axis=2)
-
-        # Add vPnvrs=0 at m=mLvlMin to close it off at the bottom (and vNvrs=0)
-        mGrid_small = np.concatenate(
-            (np.reshape(self.mLvlMinNow(self.pLvlGrid), (1, pCount)), mGrid[:, :, 0])
-        )
-        vPnvrsNow = np.concatenate(
-            (np.zeros((1, pCount)), self.u.derinv(vPnow, order=(1, 0)))
-        )
-        if self.vFuncBool:
-            vNvrsNow = np.concatenate((np.zeros((1, pCount)), self.u.inv(vNow)), axis=0)
-            vNvrsPnow = vPnow * self.u.derinv(vNow, order=(0, 1))
-            vNvrsPnow = np.concatenate((np.zeros((1, pCount)), vNvrsPnow), axis=0)
-
-        # Construct the pseudo-inverse value and marginal value functions over mLvl,pLvl
-        vPnvrsFunc_by_pLvl = []
-        vNvrsFunc_by_pLvl = []
-        for j in range(
-            pCount
-        ):  # Make a pseudo inverse marginal value function for each pLvl
-            pLvl = self.pLvlGrid[j]
-            m_temp = mGrid_small[:, j] - self.mLvlMinNow(pLvl)
-            vPnvrs_temp = vPnvrsNow[:, j]
-            vPnvrsFunc_by_pLvl.append(LinearInterp(m_temp, vPnvrs_temp))
-            if self.vFuncBool:
-                vNvrs_temp = vNvrsNow[:, j]
-                vNvrsP_temp = vNvrsPnow[:, j]
-                vNvrsFunc_by_pLvl.append(CubicInterp(m_temp, vNvrs_temp, vNvrsP_temp))
-        vPnvrsFuncBase = LinearInterpOnInterp1D(vPnvrsFunc_by_pLvl, self.pLvlGrid)
-        vPnvrsFunc = VariableLowerBoundFunc2D(
-            vPnvrsFuncBase, self.mLvlMinNow
-        )  # adjust for the lower bound of mLvl
-        if self.vFuncBool:
-            vNvrsFuncBase = LinearInterpOnInterp1D(vNvrsFunc_by_pLvl, self.pLvlGrid)
-            vNvrsFunc = VariableLowerBoundFunc2D(
-                vNvrsFuncBase, self.mLvlMinNow
-            )  # adjust for the lower bound of mLvl
-
-        # "Re-curve" the (marginal) value function
-        vPfunc = MargValueFuncCRRA(vPnvrsFunc, self.CRRA)
-        if self.vFuncBool:
-            vFunc = ValueFuncCRRA(vNvrsFunc, self.CRRA)
-        else:
-            vFunc = NullFunc()
-
-        return vFunc, vPfunc
-
-    def make_linear_xFunc(self, mLvl, pLvl, MedShk, xLvl):
-        """
-        Constructs the (unconstrained) expenditure function for this period using
-        bilinear interpolation (over permanent income and the medical shock) among
-        an array of linear interpolations over market resources.
-
-        Parameters
-        ----------
-        mLvl : np.array
-            Corresponding market resource points for interpolation.
-        pLvl : np.array
-            Corresponding permanent income level points for interpolation.
-        MedShk : np.array
-            Corresponding medical need shocks for interpolation.
-        xLvl : np.array
-            Expenditure points for interpolation, corresponding to those in mLvl,
-            pLvl, and MedShk.
-
-        Returns
-        -------
-        xFuncUnc : BilinearInterpOnInterp1D
-            Unconstrained total expenditure function for this period.
-        """
-        # Get state dimensions
-        pCount = mLvl.shape[1]
-        MedCount = mLvl.shape[0]
-
-        # Loop over each permanent income level and medical shock and make a linear xFunc
-        xFunc_by_pLvl_and_MedShk = []  # Initialize the empty list of lists of 1D xFuncs
-        for i in range(pCount):
-            temp_list = []
-            pLvl_i = pLvl[0, i, 0]
-            mLvlMin_i = self.BoroCnstNat(pLvl_i)
-            for j in range(MedCount):
-                m_temp = mLvl[j, i, :] - mLvlMin_i
-                x_temp = xLvl[j, i, :]
-                temp_list.append(LinearInterp(m_temp, x_temp))
-            xFunc_by_pLvl_and_MedShk.append(deepcopy(temp_list))
-
-        # Combine the nested list of linear xFuncs into a single function
-        pLvl_temp = pLvl[0, :, 0]
-        MedShk_temp = MedShk[:, 0, 0]
-        xFuncUncBase = BilinearInterpOnInterp1D(
-            xFunc_by_pLvl_and_MedShk, pLvl_temp, MedShk_temp
-        )
-        xFuncUnc = VariableLowerBoundFunc3D(xFuncUncBase, self.BoroCnstNat)
-        return xFuncUnc
-
-    def make_cubic_xFunc(self, mLvl, pLvl, MedShk, xLvl):
-        """
-        Constructs the (unconstrained) expenditure function for this period using
-        bilinear interpolation (over permanent income and the medical shock) among
-        an array of cubic interpolations over market resources.
-
-        Parameters
-        ----------
-        mLvl : np.array
-            Corresponding market resource points for interpolation.
-        pLvl : np.array
-            Corresponding permanent income level points for interpolation.
-        MedShk : np.array
-            Corresponding medical need shocks for interpolation.
-        xLvl : np.array
-            Expenditure points for interpolation, corresponding to those in mLvl,
-            pLvl, and MedShk.
-
-        Returns
-        -------
-        xFuncUnc : BilinearInterpOnInterp1D
-            Unconstrained total expenditure function for this period.
-        """
-        # Get state dimensions
-        pCount = mLvl.shape[1]
-        MedCount = mLvl.shape[0]
-
-        # Calculate the MPC and MPM at each gridpoint
-        EndOfPrdvPP = (
-            self.DiscFacEff
-            * self.Rfree
-            * self.Rfree
-            * np.sum(
-                self.vPPfuncNext(self.mLvlNext, self.pLvlNext) * self.ShkPrbs_temp,
-                axis=0,
-            )
-        )
-        EndOfPrdvPP = np.tile(
-            np.reshape(EndOfPrdvPP, (1, pCount, EndOfPrdvPP.shape[1])), (MedCount, 1, 1)
-        )
-        dcda = EndOfPrdvPP / self.u.der(np.array(self.cLvlNow), order=2)
-        dMedda = EndOfPrdvPP / (
-            self.MedShkVals_tiled * self.uMed.der(self.MedLvlNow, order=2)
-        )
-        dMedda[0, :, :] = 0.0  # dMedda goes crazy when MedShk=0
-        MPC = dcda / (1.0 + dcda + self.MedPrice * dMedda)
-        MPM = dMedda / (1.0 + dcda + self.MedPrice * dMedda)
-
-        # Convert to marginal propensity to spend
-        MPX = MPC + self.MedPrice * MPM
-        MPX = np.concatenate(
-            (np.reshape(MPX[:, :, 0], (MedCount, pCount, 1)), MPX), axis=2
-        )  # NEED TO CALCULATE MPM AT NATURAL BORROWING CONSTRAINT
-        MPX[0, :, 0] = self.MPCmaxNow
-
-        # Loop over each permanent income level and medical shock and make a cubic xFunc
-        xFunc_by_pLvl_and_MedShk = []  # Initialize the empty list of lists of 1D xFuncs
-        for i in range(pCount):
-            temp_list = []
-            pLvl_i = pLvl[0, i, 0]
-            mLvlMin_i = self.BoroCnstNat(pLvl_i)
-            for j in range(MedCount):
-                m_temp = mLvl[j, i, :] - mLvlMin_i
-                x_temp = xLvl[j, i, :]
-                MPX_temp = MPX[j, i, :]
-                temp_list.append(CubicInterp(m_temp, x_temp, MPX_temp))
-            xFunc_by_pLvl_and_MedShk.append(deepcopy(temp_list))
-
-        # Combine the nested list of cubic xFuncs into a single function
-        pLvl_temp = pLvl[0, :, 0]
-        MedShk_temp = MedShk[:, 0, 0]
-        xFuncUncBase = BilinearInterpOnInterp1D(
-            xFunc_by_pLvl_and_MedShk, pLvl_temp, MedShk_temp
-        )
-        xFuncUnc = VariableLowerBoundFunc3D(xFuncUncBase, self.BoroCnstNat)
-        return xFuncUnc
-
-    def make_basic_solution(self, EndOfPrdvP, aLvl, interpolator):
-        """
-        Given end of period assets and end of period marginal value, construct
-        the basic solution for this period.
-
-        Parameters
-        ----------
-        EndOfPrdvP : np.array
-            Array of end-of-period marginal values.
-        aLvl : np.array
-            Array of end-of-period asset values that yield the marginal values
-            in EndOfPrdvP.
-        interpolator : function
-            A function that constructs and returns a consumption function.
-
-        Returns
-        -------
-        solution_now : ConsumerSolution
-            The solution to this period's consumption-saving problem, with a
-            consumption function, marginal value function, and minimum m.
-        """
-        xLvl, mLvl, pLvl = self.get_points_for_interpolation(EndOfPrdvP, aLvl)
-        MedShk_temp = np.tile(
-            np.reshape(self.MedShkVals, (self.MedShkVals.size, 1, 1)),
-            (1, mLvl.shape[1], mLvl.shape[2]),
-        )
-        solution_now = self.use_points_for_interpolation(
-            xLvl, mLvl, pLvl, MedShk_temp, interpolator
-        )
-        return solution_now
-
-    def add_vPPfunc(self, solution):
-        """
-        Adds the marginal marginal value function to an existing solution, so
-        that the next solver can evaluate vPP and thus use cubic interpolation.
-
-        Parameters
-        ----------
-        solution : ConsumerSolution
-            The solution to this single period problem, which must include the
-            consumption function.
-
-        Returns
-        -------
-        solution : ConsumerSolution
-            The same solution passed as input, but with the marginal marginal
-            value function for this period added as the attribute vPPfunc.
-        """
-        vPPfuncNow = MargMargValueFuncCRRA(solution.vPfunc.cFunc, self.CRRA)
-        solution.vPPfunc = vPPfuncNow
-        return solution
-
-    def solve(self):
-        """
-        Solves a one period consumption saving problem with risky income and
-        shocks to medical need.
-
-        Parameters
-        ----------
-        None
-
-        Returns
-        -------
-        solution : ConsumerSolution
-            The solution to the one period problem, including a consumption
-            function, medical spending function ( both defined over market re-
-            sources, permanent income, and medical shock), a marginal value func-
-            tion (defined over market resources and permanent income), and human
-            wealth as a function of permanent income.
-        """
-        aLvl, trash = self.prepare_to_calc_EndOfPrdvP()
-        EndOfPrdvP = self.calc_EndOfPrdvP()
-        if self.vFuncBool:
-            self.make_EndOfPrdvFunc(EndOfPrdvP)
-        if self.CubicBool:
-            interpolator = self.make_cubic_xFunc
-        else:
-            interpolator = self.make_linear_xFunc
-        solution = self.make_basic_solution(EndOfPrdvP, aLvl, interpolator)
-        solution = self.add_MPC_and_human_wealth(solution)
-        if self.CubicBool:
-            solution = self.add_vPPfunc(solution)
-        return solution
